+++ demo/advanced/advanced.js
@@ -2,7 +2,8 @@ angular.module("demo").controller("AdvancedDemoController", function($scope) {
 
     $scope.dragoverCallback = function(event, index, external, type) {
         $scope.logListEvent('dragged over', event, index, external, type);
-        return index > 0;
+        // Disallow dropping in the third row. Could also be done with dnd-disable-if.
+        return index < 10;
     };
 
     $scope.dropCallback = function(event, index, item, external, type, allowedType) {